Improve random direction for magnetization errors
A new function is added to utils module which returns a unit vector pointing in a random direction.
The new function is used in the cassettes module for picking a random rotation axis for magnetization rotation errors in the blocks.

Previous method consisted in picking random x, y and z coordinates in a [-1,1]x[-1,1]x[-1,1] cube.
New method improves direction randomness by picking a true random 3D direction.