Fix pole positioning inconsistency
Lengths of cassette core object (core blocks or poles) are decided by the first magnetization vector in the magnetizations list (first object is pole if magnetization lies mostly in y direction).

Previously, however, at objects creation, a pole-block-pole-block sequency was always followed, starting with a pole (regardless of the previous length determination). This could lead to a mismatch between intended length and intended material for the poles/blocks, which would be hard to spot just looking at the draw() figure.
Also, such pole-block-pole-block sequency for materials also considered termination blocks, so there would be poles in the termination.

This is corrected by first creating a new "protected" attribute _is_pole_list, which is a list of booleans which are True if blocks.Block object is a pole. A related @property getter method is also created. At initialization, attribute value is [].
The list of bools is then created together with the list of length, so that the values of _is_pole_list are True for the same objects for which the length is _pole_length.
The values of is_pole_list are then directly used for testing if an object is a pole before creating the blocks.Block object.