Fix problem with numpy int positions lists
If a list of integers was previously created with numpy resulting in values of types like 'numpy.int32', these values were not recognized as int instaces, resulting in errors.
The positions list check is now performed using the numpy function for number of dimensions (not list if 0 dimensions).
(Also, a type correction).