Add log file output to model rescaling method
This commit adds the option to output the results of the model rescaling
function to a log file in JSON format. Such file contains the calculated
amplitudes and the rescaling factor (keyed by 'fres').

Also, the method is converted to a common method (not static method), so
that the field_comp attribute is also used.

The shimming documentation notebook is also updated for using the
method.

The method is renamed from get_rescale_factor to calc_rescale factor.