Add method radia_state in FildModel class
The function get_info_recursive was created in utils.py and the wrapper method for it created in
fieldsource.py, in the FieldModel class. The information obtained with the function is the geometry
type, position, dimensions, magnetization and vertices of each polyhedron that defines each division
block of some radia object.