Add flag type for tree items
QTreeWidgetItem subclassed as ExploreItem and added item_type argument for identification of what the item. The first types constructed in Items class are Data, Model, Field and Trajectory;
Added warning messages to alert of the reload of files and recalculation of trajectory;
Update AnalysisPushButton to uncheck the items every time the menu is open.