Great update over almost all of the scripts
In the application file, used time module for count the import time intervals of the blocks of other modules;
Created "Apply for All" action in the Settings menu to apply the checked analysis for all the data loaded in the current project widget displayed;
Style of table and TableView instances moved to table_model.py;
In ModelDialog class, the code of model parameters group box was moved for DeltaLayout. Included Delta Prototype and Delta Carnauba models in ModelDialog and created method for export the data configurate in the spin boxes;
Accept answer of the model dialog creates label referent to the model generated in the project tree;
Debbuged the Analysis button + menu;
Included tree item type for the top level items Data and Models.