Implement missing analysis
Complemented the set of enums of the analysis items in ExploreItem. Created subclass of QListWidgetItem to add the superior and subordinate parameters;
In the analysis button class, shimming was excluded and kickmap too. Corrected bugs in the button menu items behavior. Implemented the relation between superior and subordinate list items in the menu;
Added the remaining dictionaries for analysis in ProjectWidget and created one more column in the project tree for description or presentation of each tree item content;
Finally, in the MainWindow, coded the missing analysis methods.