Table features, toolbar selection mode and icons
Close window using the close button (X) in the top right corner now leads to the same quit confirmation dialog as the "Exit" action of the File menu;
Icons added to menu actions in menu bar;
Action "Cursor" inserted in Tool bar. Buttons in tool bar check and uncheck themselves when user click in one or another;
Table class accepts data only in the form of numpy array, not InsertionDeviceData anymore. Table button now can display the content of "List" tree items. It was also implemented a option in the Plot button menu to plot selected columns of tables.