Canvas lines and legend edit
matplotlib 3.7.2 default figure options and _formlayout modified to edit also:
-Axes: title size, label size, tick size and grid (on/off);
-Curves: include legend of curve or not;
-Legend: display, drag, frame, shadow, fancy box, font size and alpha.

Created Figure Options dock widget in right side of main window to expose the tab widget to edit the lines and legends of the plots;
Created methods in VisualizationTabWidget to deal with open of multiple figure options from plot tabs. Methods made to achieve the requiered behaviour of close figure options when visualization window tab is changed or closed and don't have multiple figure options widgets into the Figure Options dock widget.

Set layout of figure in Canvas to tight, then after each draw, the figure is automatically updated to tight_layout;
Changed plots and draw methods of VisualizationTabWidget and ProjectWidget to set directly the labels into the argument of ax.plot(), substituting the return of the lines and labels.

Added statustip to exhibit the path of the data loaded. This tip can be displayed or not by checking option in the View menu.