New analysis: Integrals vs X
Warning dialog of repeated analysis tried to be done changed to Question dialog. The user is asked for replace the last same analysis done or add a new one or just ignore and don't calculate it. Details: the LAST analysis of same type is replaced, if this option was selected; Phase Error is dependent of Trajectory, then the phase error is calculated for the last trajectory done;
Added new analysis of Total Field Integrals calculated for a range of x horizontal positions with its own range of x and z and y value.