Extract Field Map of ID Model
A new action/button was added in the Tool Bar to create a field map from a model. Trigger the action and select the ID model item, then specify the grid and confirm Ok in the ModelData Dialog;
The range limits of nr_periods, period_length and gap were improved in the data import dialog.