Operations between Items
New button "Operations" was added to the tool bar. It has the options "plus" and "minus" to sum or subtract two or more data from analysis or result items;
Behaviour: while the button is active, select many or analysis items of same kind or result items compatible (same array length), then press spacebar to operate them. A new dock window will appear with the containers of custom analyses and results items and the result of the operation. The address of the parent items used in the operation can be found in the status bar while the mouse cursor is over the operation result item;

The OperationsDockWidget, where the result operation items live, can be hid or showed in the View Menu;
Plot and Table work as usual in the result operation items;
Plot errors now is presented not as text in terminal, but as QMessageBox dialogs. This was also implemented to Operation errors;
Configuration of ExploreWindowTreeWidget was separated to new base class BasicTreeWidget. This basic tree also was used to configurate the OperationTreeWidget.