Correct plot bugs and others improvements
Plot validation improved, then the bugs were corrected: try to plot two different analyses at same time; try to plot pair of arrays of different sizes.
Apply for All mode for analysis now is indicated by icon in analysis button when this option is active and "Apply" button of analysis menu is rewritten to "Apply for All".
Cursor shortcut changed from "C" key to "Esc" key.
Cat, dog and bug options removed from Table button menu.
Label name numeration corrected. Now the initial labels are stored in dict without the numbers to count the repeated ones. A dict was used to allow the remotion of the label referent to some ID item when this item is deleted.
Range of number of periods and of the others parameters set for models.
Question message inserted to confirm the close of some project tab.