Solve radia model and SummaryWidget optimized
Added new button to tool bar, the SolveModel button. When activated and a model is clicked, the radia object relaxation procedure is executed. The method of solve is located at the ProjectWidget class and connected into the MainWindow. Also, when the solve is done, the model icon earns a tick to indicate that it is solved;
treeItemInfo method now stores the container item;
AnalysisDialog class name changed to EditAnalysisDialog; AnalysisPushButton class name changed to AnalysisButton;
Tab order set in DataDialog to move focus from line to line, after last line in the grid, the focus is moved to the right column;
Code of SummaryWidgets, its conditions in MainWindow and the update of IDs, phase error and integrals were improved;